+++ doorstop/core/types.py
@@ -15,6 +15,8 @@ class Prefix(str):  # pylint: disable=R0904
 
     """Unique document prefixes."""
 
+    UNKNOWN_MESSGE = "no document with prefix: {}"
+
     def __new__(cls, value=""):
         if isinstance(value, Prefix):
             return value
@@ -54,6 +56,8 @@ class Prefix(str):  # pylint: disable=R0904
 class ID(object):
 
     """Unique item identifier."""
+
+    UNKNOWN_MESSAGE = "no{k} item with ID: {i}"  # k='parent'|'child'|'', i=ID
 
     def __new__(cls, *values):
         if values and isinstance(values[0], ID):
@@ -128,16 +132,19 @@ class ID(object):
     @property
     def prefix(self):
         """Get the ID's prefix."""
-        if self._exc:
-            raise self._exc
+        self.check()
         return self._prefix
 
     @property
     def number(self):
         """Get the ID's number."""
+        self.check()
+        return self._number
+
+    def check(self):
+        """Verify an ID is valid."""
         if self._exc:
             raise self._exc
-        return self._number
 
     @staticmethod
     def split_id(text):
@@ -195,6 +202,7 @@ class Text(str):  # pylint: disable=R0904
     """Markdown text paragraph."""
 
     def __new__(cls, value=""):
+        assert not isinstance(value, Text)
         obj = super(Text, cls).__new__(cls, Text.load_text(value))
         return obj
 
@@ -219,7 +227,7 @@ class Text(str):  # pylint: disable=R0904
     @staticmethod
     def save_text(text, end='\n'):
         """Break a string at sentences and dump as wrapped literal YAML."""
-        return Literal(Text.wrap(Text.sbd(text, end=end)))
+        return Literal(Text.wrap(Text.sbd(str(text), end=end)))
 
     # Based on: http://en.wikipedia.org/wiki/Sentence_boundary_disambiguation
     RE_SENTENCE_BOUNDARIES = re.compile(r"""
@@ -343,7 +351,8 @@ class Level(object):
                 self._parts = parts
                 self.heading = False
         self.heading = self.heading if heading is None else heading
-        self._adjust()
+        if not value:
+            self._adjust()
 
     def __repr__(self):
         if self.heading:
@@ -534,14 +543,33 @@ class Level(object):
         return Level(self.value)
 
 
+def is_tree(obj):
+    """Determine if the object is a tree-like."""
+    return hasattr(obj, 'documents')
+
+
 def is_document(obj):
-    """Determine if the object is a document."""
+    """Determine if the object is a document-like."""
     return hasattr(obj, 'items')
 
 
-def is_tree(obj):
-    """Determine if the object is a tree."""
-    return hasattr(obj, 'documents')
+def is_item(obj):
+    """Determine if the object is item-like."""
+    return hasattr(obj, 'text')
+
+
+def iter_documents(obj, path, ext):
+    """Get an iterator if documents from a tree or document-like object."""
+    if is_tree(obj):
+        # a tree
+        logging.debug("iterating over tree...")
+        for document in obj.documents:
+            path2 = os.path.join(path, document.prefix + ext)
+            yield document, path2
+    else:
+        # assume a document-like object
+        logging.debug("iterating over document-like object...")
+        yield obj, path
 
 
 def iter_items(obj):
@@ -558,17 +586,3 @@ def iter_items(obj):
         # an item
         logging.debug("iterating over an item (in a container)...")
         return [obj]
-
-
-def iter_documents(obj, path, ext):
-    """Get an iterator if documents from a tree or document-like object."""
-    if is_tree(obj):
-        # a tree
-        logging.debug("iterating over tree...")
-        for document in obj.documents:
-            path2 = os.path.join(path, document.prefix + ext)
-            yield document, path2
-    else:
-        # assume a document-like object
-        logging.debug("iterating over document-like object...")
-        yield obj, path